requirements.txt
dropbox
openpyxl
click
colorama
loguru
numpy
tabulate
python-dotenv
